fix: updated missing migration file
