+++ alembic/versions/0d98e6333de6_update_contact_us_model_with_org_id.py
@@ -1,0 +1,55 @@
+"""update contact-us model with org_id
+
+Revision ID: 0d98e6333de6
+Revises: 5f9d01dc6cc9
+Create Date: 2024-07-31 08:52:20.606379
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '0d98e6333de6'
+down_revision: Union[str, None] = '5f9d01dc6cc9'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('notification_settings',
+    sa.Column('mobile_push_notifications', sa.Boolean(), server_default='false', nullable=True),
+    sa.Column('email_notification_activity_in_workspace', sa.Boolean(), server_default='false', nullable=True),
+    sa.Column('email_notification_always_send_email_notifications', sa.Boolean(), server_default='false', nullable=True),
+    sa.Column('email_notification_email_digest', sa.Boolean(), server_default='false', nullable=True),
+    sa.Column('email_notification_announcement_and_update_emails', sa.Boolean(), server_default='false', nullable=True),
+    sa.Column('slack_notifications_activity_on_your_workspace', sa.Boolean(), server_default='false', nullable=True),
+    sa.Column('slack_notifications_always_send_email_notifications', sa.Boolean(), server_default='false', nullable=True),
+    sa.Column('slack_notifications_announcement_and_update_emails', sa.Boolean(), server_default='false', nullable=True),
+    sa.Column('user_id', sa.String(), nullable=False),
+    sa.Column('id', sa.String(), nullable=False),
+    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
+    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
+    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_notification_settings_id'), 'notification_settings', ['id'], unique=False)
+    op.add_column('contact_us', sa.Column('phone_number', sa.String(), nullable=False))
+    op.add_column('contact_us', sa.Column('org_id', sa.String(), nullable=False))
+    op.create_foreign_key(None, 'contact_us', 'organizations', ['org_id'], ['id'])
+    op.drop_column('contact_us', 'title')
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column('contact_us', sa.Column('title', sa.VARCHAR(), autoincrement=False, nullable=False))
+    op.drop_constraint(None, 'contact_us', type_='foreignkey')
+    op.drop_column('contact_us', 'org_id')
+    op.drop_column('contact_us', 'phone_number')
+    op.drop_index(op.f('ix_notification_settings_id'), table_name='notification_settings')
+    op.drop_table('notification_settings')
+    # ### end Alembic commands ###
